requirements.txt
numpy==1.20.3
Pillow==9.0.0
streamlit==1.5.0
